Update EightPuzzleGame_UinformedSearch.py
Uninformed search is now working!
Removed or commented out many debugging statements
Cleaned up the function that searched the closed list and compared the stated created by swap to see if it is an already visited state.
Fixed a tabbing issue in the if statements that preformed swaps.  Swap Right could only execute if a swap down happened first because it was tabbed into the swap down if, correcting this fixed uninformed search.  Huzzah the project is 30% complete.