h2 heuristic, distance out of place
Added the code for the h2 heuristic sum of distance out of place.  It finds the x and y coordinates of the out of place tiles, then finds the x and y coordinates of the correct location, and returns the absolute value of their sum.  It was actually pretty easy to do, if a bit hard to read because of all the letter variables.