Update EightPuzzleGame_UinformedSearch.py
Cleaned up the uninformed search, did the for loops different so they had a built in iterator instead of me trying to make a separate iterator, removed all of the commented out code that did not work, and removed all of the debugging statements.  Also implemented depth tracking to find the length of the path.